Update ChessBoard.cpp
Reworked DisplayBoard (not very much different though), to display the board whenever called. initializeGrid is called upon the creation of chessBoard to intialize the pieces on the chess board and the chess board grid node labels themselves. Additionally, the currGridNode for the chess piece is also set to be at the initial state.