Update ChessBoard.h
Function definitions commented out are yet to be implemented and the plan of the functionality is listed as comments above it. Right now, the chess board is reworked as being a 2d array of grid_node structs, where these structs are initialized with their values in the initializeGrid() function that is called within the chessBoard constructor. DeterminePossibleMoves tries to find the legal moves of pieces in many situations given the current location of the chess piece (exceptions are it doesnt do it for king pieces, and pawns are not allowed to move 2 spaces at the start yet, also en passant is not implemented).