Update ChessPiece.h
Big change made was to not utilize inheritance and to consider all pieces to be implemented under one class, chessPiece, the attributes serve as the identifiers for the identity of each piece. The unique identifier for each piece is currNodeLabel (current location of a given piece on the board which is unique for every piece).